benchmarks

+++ read_stride/read_stride.cpp
@@ -22,7 +22,7 @@ static void BM_read_stride_128kb(benchmark::State& state) {
 	free(buffer);
 
 	// Bytes written = # Iterations * bytes written per iteration
-         state.counters["IO Bytes"] = (MAX_BYTES / (state.range(0) + read_0)) * read_0;
+         state.counters["IO Bytes"] = (MAX_BYTES / (double)(state.range(0) + read_0)) * read_0;
 }
 
 BENCHMARK(BM_read_stride_128kb)->RangeMultiplier(2)->Range(1<<12, 1<<27);
