+++ read_stride/read_stride_1.cpp
@@ -23,7 +23,7 @@ static void BM_read_stride_256kb(benchmark::State& state) {
 	free(buffer);
 
 	// Bytes written = # Iterations * bytes written per iteration
-         state.counters["IO Bytes"] = (MAX_BYTES / (state.range(0) + read_1)) * read_1;
+         state.counters["IO Bytes"] = (MAX_BYTES / (double)(state.range(0) + read_1)) * read_1;
 }
 
 BENCHMARK(BM_read_stride_256kb)->RangeMultiplier(2)->Range(1<<12, 1<<27);
