+++ read_stride/read_stride_4.cpp
@@ -23,7 +23,7 @@ static void BM_read_stride_4MB(benchmark::State& state) {
 	free(buffer);
 
 	// Bytes written = # Iterations * bytes written per iteration
-         state.counters["IO Bytes"] = (MAX_BYTES / (state.range(0) + read_4)) * read_4;
+         state.counters["IO Bytes"] = (MAX_BYTES / (double)(state.range(0) + read_4)) * read_4;
 }
 
 BENCHMARK(BM_read_stride_4MB)->RangeMultiplier(2)->Range(1<<12, 1<<27);
