+++ write_stride/write_stride.cpp
@@ -23,7 +23,7 @@ static void BM_write_stride_128kb(benchmark::State& state) {
 	free(buffer);
 
 	// Bytes written = # Iterations * bytes written per iteration
-	state.counters["IO Bytes"] = (MAX_BYTES / (state.range(0) + write_0)) * write_0;
+	state.counters["IO Bytes"] = (MAX_BYTES / (double)(state.range(0) + write_0)) * write_0;
 }
 
 BENCHMARK(BM_write_stride_128kb)->RangeMultiplier(2)->Range(1<<12, 1<<27);
