+++ write_stride/write_stride_1.cpp
@@ -21,7 +21,7 @@ static void BM_write_stride_256kb(benchmark::State& state) {
 	// free buffer
 	free(buffer);
 	// Bytes written = # Iterations * bytes written per iteration
-        state.counters["IO Bytes"] = (MAX_BYTES / (state.range(0) + write_1)) * write_1;
+        state.counters["IO Bytes"] = (MAX_BYTES / (double)(state.range(0) + write_1)) * write_1;
 }
 
 BENCHMARK(BM_write_stride_256kb)->RangeMultiplier(2)->Range(1<<12, 1<<27);
