+++ write_stride/write_stride_2.cpp
@@ -21,7 +21,7 @@ static void BM_write_stride_1MB(benchmark::State& state) {
 	free(buffer);
 
 	 // Bytes written = # Iterations * bytes written per iteration
-         state.counters["IO Bytes"] = (MAX_BYTES / (state.range(0) + write_2)) * write_2;
+         state.counters["IO Bytes"] = (MAX_BYTES / (double)(state.range(0) + write_2)) * write_2;
 }
 
 BENCHMARK(BM_write_stride_1MB)->RangeMultiplier(2)->Range(1<<12, 1<<27);
