+++ write_stride/write_stride_3.cpp
@@ -22,7 +22,7 @@ static void BM_write_stride_2MB(benchmark::State& state) {
 	free(buffer);
 
 	// Bytes written = # Iterations * bytes written per iteration
-         state.counters["IO Bytes"] = (MAX_BYTES / (state.range(0) + write_3)) * write_3;
+         state.counters["IO Bytes"] = (MAX_BYTES / (double)(state.range(0) + write_3)) * write_3;
 }
 
 BENCHMARK(BM_write_stride_2MB)->RangeMultiplier(2)->Range(1<<12, 1<<27);
